refactor: 피드백 반영
- NavMenuContentStyle css function 추가
- /components/@layout으로 폴더 네이밍 수정
- /pages/home, /pages/search 로 페이지 컴포넌트 위치 수정
